requirements.txt
streamlit>=1.30.0
pandas
requests
sqlalchemy
psycopg2-binary
python-dotenv
altair<5.0.0